backend/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-groq
pydantic
python-dotenv
